requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
joblib==1.4.2
nltk==3.9.1
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytesseract==0.3.13
qrcode==8.0
regex==2024.11.6
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.1
gunicorn